+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - BREAKING: Added CLI `watch_namespace` parameter to ProductOperatorRun in
-  preparation for operators watching a single namespace ([#332]).
+  preparation for operators watching a single namespace ([#332], [#333]).
 - More builder functionality ([#331]) 
   - builder for `SecurityContext` objects 
   - add `EnvVar`s from field refs
@@ -17,7 +17,6 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Build against Kubernetes 1.23 ([#330]).
-- BREAKING: Using `WatchNamespace` now for CLI `watch_namespace` ([#333]).
 
 [#330]: https://github.com/stackabletech/operator-rs/pull/330
 [#331]: https://github.com/stackabletech/operator-rs/pull/331